vogon/requirements.txt
conceptpower-api==1.0
Django==1.8.2
djangorestframework==3.1.3
requests==2.7.0
